Move models to a separate file
Unsure if this is the convention the repo will go with, or if there will
be a global types.ts file instead for whole app
